updated tests and split them out
