+++ python/pytests/golden/count_test/test_count_windowed.jsonl
