+++ python/pytests/golden/sum_test/test_sum_unwindowed.jsonl
